requirements.txt
discohook
gspread
oauth2client
python-dotenv
aiohttp
